requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
rdkit-pypi>=2022.03.1
chembl-webresource-client>=0.10.0
lazypredict>=0.2.12
padelpy>=0.1.11
joblib>=1.1.0
scipy>=1.7.0
requests>=2.25.0
Pillow>=9.0.0